+++ src/components/account/Account.js
@@ -719,6 +719,7 @@ class Account extends Component {
     const user_info = store.user_info || {wallets: []};
     return (
       <View
+        key="wallets"
         style={{
           // marginTop: 7,
           borderTopWidth: 0,
@@ -764,7 +765,7 @@ class Account extends Component {
   renderExtWallets() {
     const user_info = store.user_info || {ext_wallets: []};
     return (
-      <View>
+      <View key="extra_wallets">
         <View style={styles.add_store_actions_box}>
           {user_info.ext_wallets.map((wallet, index) => (
             <TouchableHighlight