+++ src/containers/Home/Home.js
@@ -38,6 +38,7 @@ class Home extends Component {
       },
       primaryActions: null,
       product_groups: [],
+      news_categories: [],
     };
     this.eventTracker = new EventTracker();
   }
@@ -112,6 +113,7 @@ class Home extends Component {
           primaryActions: response.data.primary_actions,
           showPrimaryActions: response.data.showPrimaryActions,
           product_groups: response.data.product_groups,
+          news_categories: response.data.news_categories,
         }));
 
         this.executeDeepLink();
@@ -196,7 +198,7 @@ class Home extends Component {
 
   handlePressService(service, callBack) {
     const {t} = this.props;
-    
+
     if (service.type === 'chat') {
       this.handlePressButtonChat(this.state.site);
     } else {
@@ -224,8 +226,11 @@ class Home extends Component {
     });
   };
 
-  handleShowAllNews = () => {
-    Actions.push(appConfig.routes.newsTab);
+  handleShowAllNews = (title, id) => {
+    Actions.push(appConfig.routes.newsTab, {
+      title,
+      id,
+    });
   };
 
   handlePressRoomItem = room => {
@@ -312,7 +317,9 @@ class Home extends Component {
         promotions={this.state.promotions}
         listService={this.state.listService}
         listServiceType={this.state.listServiceConfig.type}
-        listServiceItemsPerRow={Number(this.state.listServiceConfig.items_per_row)}
+        listServiceItemsPerRow={Number(
+          this.state.listServiceConfig.items_per_row,
+        )}
         primaryActions={this.state.primaryActions}
         showPrimaryActions={this.state.showPrimaryActions}
         apiFetching={this.state.apiFetching}
@@ -327,7 +334,7 @@ class Home extends Component {
         onShowAllGroupProduct={this.handleShowAllGroupProduct}
         onShowAllSites={this.handleShowAllSites}
         onShowAllCampaigns={this.handleShowAllCampaigns}
-        onShowAllNews={this.handleShowAllNews}
+        onShowAllNews={(title, id) => this.handleShowAllNews(title, id)}
         onPressSiteItem={this.handlePressSiteItem}
         onPressRoomItem={this.handlePressRoomItem}
         onPressCampaignItem={this.handlePressCampaignItem}
@@ -336,6 +343,8 @@ class Home extends Component {
         onPressNoti={this.handlePressButtonChat}
         refreshing={this.state.refreshing}
         groups={this.state.product_groups}
+        product_groups={this.state.product_groups}
+        news_categories={this.state.news_categories}
       />
     );
   }